pyproject.toml
[project]
name = "lol-rag-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "gliner>=0.2.26",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "langchain-postgres>=0.0.17",
    "matplotlib>=3.10.8",
    "neo4j>=6.1.0",
    "networkx>=3.6.1",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "seaborn>=0.13.2",
    "torch>=2.11.0",
    "transformers>=5.1.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
